Webでデプロイしたページに何も表示されない不具合を修正するため、環境変数の管理方式を変更
