+++ .github/workflows/flutter-web-cd.yml
@@ -43,7 +43,7 @@ jobs:
 
       - name: Build Flutter Web
         working-directory: app
-        run: flutter build web --release --base-href /
+        run: flutter build web --release --base-href / -t lib/main_${{ matrix.environment }}.dart
 
       - name: Configure AWS Credentials
         uses: aws-actions/configure-aws-credentials@v4
