+++ app/lib/main.dart
@@ -1,26 +1,2 @@
-import 'package:app/config/amplify_initializer.dart';
-import 'package:app/screens/sample_screen/sample_screen.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_riverpod/flutter_riverpod.dart';
-
-Future<void> main() async {
-  WidgetsFlutterBinding.ensureInitialized();
-  await configureAmplify();
-  runApp(const ProviderScope(child: MyApp()));
-}
-
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Genkaimeshi Recipe',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
-      ),
-      home: const SampleScreen(),
-    );
-  }
-}
+// デフォルトはdev環境で起動
+export 'main_dev.dart';
